+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [fix:remove the secondary report.](https://github.com/Tencent/spring-cloud-tencent/pull/901)
 - [fix:optimize instance circuit beaker.](https://github.com/Tencent/spring-cloud-tencent/pull/910)
 - [fix:optimize multi service registration and discovery.](https://github.com/Tencent/spring-cloud-tencent/pull/915)
+- [feature: improve circuit breaker usage.](https://github.com/Tencent/spring-cloud-tencent/pull/917)